Safari-609.1.11.1/LayoutTests/svg/animations/animate-end-attribute-numeric-precision-expected.txt

SVG 1.1 dynamic animation tests

Tests end conditions are respected properly near the limits of float numeric precision

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 300
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
